Clean params classes
