+++ pygem/__init__.py
@@ -1,9 +1,10 @@
 """
 """
 __all__ = [
-    'affine', 'filehandler', 'freeform', 'radial', 'openfhandler', 'params',
+    'affine', 'filehandler', 'freeform', 'radial', 'openfhandler',
     'stlhandler', 'unvhandler', 'vtkhandler', 'nurbshandler', 'stephandler',
-    'igeshandler', 'utils', 'gui', 'khandler', 'idw', 'params_idw'
+    'igeshandler', 'utils', 'gui', 'khandler', 'idw', 'params_idw',
+    'params_rbf', 'params_ffd'
 ]
 
 from . import affine
@@ -11,7 +12,6 @@ from . import filehandler
 from . import freeform
 from . import radial
 from . import openfhandler
-from . import params
 from . import stlhandler
 from . import unvhandler
 from . import vtkhandler
@@ -23,3 +23,5 @@ from . import gui
 from . import khandler
 from . import idw
 from . import params_idw
+from . import params_rbf
+from . import params_ffd
