+++ tests/test_ffdparams.py
@@ -7,127 +7,127 @@ from OCC.BRepAlgoAPI import BRepAlgoAPI_Cut
 from OCC.BRepPrimAPI import BRepPrimAPI_MakeSphere, BRepPrimAPI_MakeBox
 from OCC.gp import gp_Pnt
 
-import pygem.params as ffdp
+from pygem.params_ffd import FFDParameters
 
 
 class TestFFDParameters(TestCase):
     def test_class_members_default_n_control_points(self):
-        params = ffdp.FFDParameters()
+        params = FFDParameters()
         assert np.array_equal(params.n_control_points, [2, 2, 2])
 
     def test_class_members_default_conversion_unit(self):
-        params = ffdp.FFDParameters()
+        params = FFDParameters()
         assert params.conversion_unit == 1.
 
     def test_class_members_default_lenght_box(self):
-        params = ffdp.FFDParameters()
+        params = FFDParameters()
         assert np.array_equal(params.lenght_box, np.ones(3))
 
     def test_class_members_default_origin_box(self):
-        params = ffdp.FFDParameters()
+        params = FFDParameters()
         assert np.array_equal(params.origin_box, np.zeros(3))
 
     def test_class_members_default_rot_angle(self):
-        params = ffdp.FFDParameters()
+        params = FFDParameters()
         assert np.array_equal(params.rot_angle, np.zeros(3))
 
     def test_class_members_default_array_mu_x(self):
-        params = ffdp.FFDParameters()
+        params = FFDParameters()
         np.testing.assert_array_almost_equal(params.array_mu_x,
                                              np.zeros((2, 2, 2)))
 
     def test_class_members_default_array_mu_y(self):
-        params = ffdp.FFDParameters()
+        params = FFDParameters()
         np.testing.assert_array_almost_equal(params.array_mu_y,
                                              np.zeros((2, 2, 2)))
 
     def test_class_members_default_array_mu_z(self):
-        params = ffdp.FFDParameters()
+        params = FFDParameters()
         np.testing.assert_array_almost_equal(params.array_mu_z,
                                              np.zeros((2, 2, 2)))
 
     def test_class_members_default_psi_mapping(self):
-        params = ffdp.FFDParameters()
+        params = FFDParameters()
         np.testing.assert_array_almost_equal(params.psi_mapping,
                                              np.diag([1, 1, 1]))
 
     def test_class_members_default_inv_psi_mapping(self):
-        params = ffdp.FFDParameters()
+        params = FFDParameters()
         np.testing.assert_array_almost_equal(params.inv_psi_mapping,
                                              np.diag([1, 1, 1]))
 
     def test_class_members_default_rotation_matrix(self):
-        params = ffdp.FFDParameters()
+        params = FFDParameters()
         np.testing.assert_array_almost_equal(params.rotation_matrix, np.eye(3))
 
     def test_class_members_default_position_vertex_0(self):
-        params = ffdp.FFDParameters()
+        params = FFDParameters()
         np.testing.assert_array_almost_equal(params.position_vertex_0,
                                              np.zeros(3))
 
     def test_class_members_default_position_vertex_1(self):
-        params = ffdp.FFDParameters()
+        params = FFDParameters()
         np.testing.assert_array_almost_equal(params.position_vertex_1,
                                              np.array([1., 0., 0.]))
 
     def test_class_members_default_position_vertex_2(self):
-        params = ffdp.FFDParameters()
+        params = FFDParameters()
         np.testing.assert_array_almost_equal(params.position_vertex_2,
                                              np.array([0., 1., 0.]))
 
     def test_class_members_default_position_vertex_3(self):
-        params = ffdp.FFDParameters()
+        params = FFDParameters()
         np.testing.assert_array_almost_equal(params.position_vertex_3,
                                              np.array([0., 0., 1.]))
 
     def test_class_members_generic_n_control_points(self):
-        params = ffdp.FFDParameters([2, 3, 5])
+        params = FFDParameters([2, 3, 5])
         assert np.array_equal(params.n_control_points, [2, 3, 5])
 
     def test_class_members_generic_array_mu_x(self):
-        params = ffdp.FFDParameters([2, 3, 5])
+        params = FFDParameters([2, 3, 5])
         np.testing.assert_array_almost_equal(params.array_mu_x,
                                              np.zeros((2, 3, 5)))
 
     def test_class_members_generic_array_mu_y(self):
-        params = ffdp.FFDParameters([2, 3, 5])
+        params = FFDParameters([2, 3, 5])
         np.testing.assert_array_almost_equal(params.array_mu_y,
                                              np.zeros((2, 3, 5)))
 
     def test_class_members_generic_array_mu_z(self):
-        params = ffdp.FFDParameters([2, 3, 5])
+        params = FFDParameters([2, 3, 5])
         np.testing.assert_array_almost_equal(params.array_mu_z,
                                              np.zeros((2, 3, 5)))
 
     def test_read_parameters_conversion_unit(self):
-        params = ffdp.FFDParameters(n_control_points=[3, 2, 2])
+        params = FFDParameters(n_control_points=[3, 2, 2])
         params.read_parameters('tests/test_datasets/parameters_sphere.prm')
         assert params.conversion_unit == 1.
 
     def test_read_parameters_n_control_points(self):
-        params = ffdp.FFDParameters(n_control_points=[3, 2, 2])
+        params = FFDParameters(n_control_points=[3, 2, 2])
         params.read_parameters('tests/test_datasets/parameters_sphere.prm')
         assert np.array_equal(params.n_control_points, [3, 2, 2])
 
     def test_read_parameters_lenght_box_x(self):
-        params = ffdp.FFDParameters(n_control_points=[3, 2, 2])
+        params = FFDParameters(n_control_points=[3, 2, 2])
         params.read_parameters('tests/test_datasets/parameters_sphere.prm')
         assert np.array_equal(params.lenght_box, [45.0, 90.0, 90.0])
 
     def test_read_parameters_origin_box(self):
-        params = ffdp.FFDParameters(n_control_points=[3, 2, 2])
+        params = FFDParameters(n_control_points=[3, 2, 2])
         params.read_parameters('tests/test_datasets/parameters_sphere.prm')
         origin_box_exact = np.array([-20.0, -55.0, -45.0])
         np.testing.assert_array_almost_equal(params.origin_box,
                                              origin_box_exact)
 
     def test_read_parameters_rot_angle_x(self):
-        params = ffdp.FFDParameters(n_control_points=[3, 2, 2])
+        params = FFDParameters(n_control_points=[3, 2, 2])
         params.read_parameters('tests/test_datasets/parameters_sphere.prm')
         assert np.array_equal(params.rot_angle, [20.3, 11.0, 0.])
 
     def test_read_parameters_array_mu_x(self):
-        params = ffdp.FFDParameters(n_control_points=[3, 2, 2])
+        params = FFDParameters(n_control_points=[3, 2, 2])
         params.read_parameters('tests/test_datasets/parameters_sphere.prm')
         array_mu_x_exact = np.array(
             [0.2, 0., 0., 0., 0.5, 0., 0., 0., 1., 0., 0., 0.]).reshape((3, 2,
@@ -137,7 +137,7 @@ class TestFFDParameters(TestCase):
                                              array_mu_x_exact)
 
     def test_read_parameters_array_mu_y(self):
-        params = ffdp.FFDParameters(n_control_points=[3, 2, 2])
+        params = FFDParameters(n_control_points=[3, 2, 2])
         params.read_parameters('tests/test_datasets/parameters_sphere.prm')
         array_mu_y_exact = np.array(
             [0., 0., 0.5555555555, 0., 0., 0., 0., 0., -1., 0., 0.,
@@ -146,7 +146,7 @@ class TestFFDParameters(TestCase):
                                              array_mu_y_exact)
 
     def test_read_parameters_array_mu_z(self):
-        params = ffdp.FFDParameters(n_control_points=[3, 2, 2])
+        params = FFDParameters(n_control_points=[3, 2, 2])
         params.read_parameters('tests/test_datasets/parameters_sphere.prm')
         array_mu_z_exact = np.array(
             [0., -0.2, 0., -0.45622985, 0., 0., 0., 0., -1.22, 0., -1.,
@@ -155,21 +155,21 @@ class TestFFDParameters(TestCase):
                                              array_mu_z_exact)
 
     def test_read_parameters_psi_mapping(self):
-        params = ffdp.FFDParameters(n_control_points=[3, 2, 2])
+        params = FFDParameters(n_control_points=[3, 2, 2])
         params.read_parameters('tests/test_datasets/parameters_sphere.prm')
         psi_mapping_exact = np.diag([0.02222222, 0.01111111, 0.01111111])
         np.testing.assert_array_almost_equal(params.psi_mapping,
                                              psi_mapping_exact)
 
     def test_read_parameters_inv_psi_mapping(self):
-        params = ffdp.FFDParameters(n_control_points=[3, 2, 2])
+        params = FFDParameters(n_control_points=[3, 2, 2])
         params.read_parameters('tests/test_datasets/parameters_sphere.prm')
         inv_psi_mapping_exact = np.diag([45., 90., 90.])
         np.testing.assert_array_almost_equal(params.inv_psi_mapping,
                                              inv_psi_mapping_exact)
 
     def test_read_parameters_rotation_matrix(self):
-        params = ffdp.FFDParameters(n_control_points=[3, 2, 2])
+        params = FFDParameters(n_control_points=[3, 2, 2])
         params.read_parameters('tests/test_datasets/parameters_sphere.prm')
         rotation_matrix_exact = np.array(
             [[0.98162718, 0., 0.190809], [0.06619844, 0.93788893, -0.34056147],
@@ -178,20 +178,20 @@ class TestFFDParameters(TestCase):
                                              rotation_matrix_exact)
 
     def test_read_parameters_position_vertex_0_origin(self):
-        params = ffdp.FFDParameters(n_control_points=[3, 2, 2])
+        params = FFDParameters(n_control_points=[3, 2, 2])
         params.read_parameters('tests/test_datasets/parameters_sphere.prm')
         np.testing.assert_array_almost_equal(params.position_vertex_0,
                                              params.origin_box)
 
     def test_read_parameters_position_vertex_0(self):
-        params = ffdp.FFDParameters(n_control_points=[3, 2, 2])
+        params = FFDParameters(n_control_points=[3, 2, 2])
         params.read_parameters('tests/test_datasets/parameters_sphere.prm')
         position_vertex_0_exact = np.array([-20.0, -55.0, -45.0])
         np.testing.assert_array_almost_equal(params.position_vertex_0,
                                              position_vertex_0_exact)
 
     def test_read_parameters_position_vertex_1(self):
-        params = ffdp.FFDParameters(n_control_points=[3, 2, 2])
+        params = FFDParameters(n_control_points=[3, 2, 2])
         params.read_parameters('tests/test_datasets/parameters_sphere.prm')
         position_vertex_1_exact = np.array(
             [24.17322326, -52.02107006, -53.05309404])
@@ -199,14 +199,14 @@ class TestFFDParameters(TestCase):
                                              position_vertex_1_exact)
 
     def test_read_parameters_position_vertex_2(self):
-        params = ffdp.FFDParameters(n_control_points=[3, 2, 2])
+        params = FFDParameters(n_control_points=[3, 2, 2])
         params.read_parameters('tests/test_datasets/parameters_sphere.prm')
         position_vertex_2_exact = np.array([-20., 29.41000412, -13.77579136])
         np.testing.assert_array_almost_equal(params.position_vertex_2,
                                              position_vertex_2_exact)
 
     def test_read_parameters_position_vertex_3(self):
-        params = ffdp.FFDParameters(n_control_points=[3, 2, 2])
+        params = FFDParameters(n_control_points=[3, 2, 2])
         params.read_parameters('tests/test_datasets/parameters_sphere.prm')
         position_vertex_3_exact = np.array(
             [-2.82719042, -85.65053198, 37.85915459])
@@ -214,19 +214,19 @@ class TestFFDParameters(TestCase):
                                              position_vertex_3_exact)
 
     def test_read_parameters_failing_filename_type(self):
-        params = ffdp.FFDParameters(n_control_points=[3, 2, 2])
+        params = FFDParameters(n_control_points=[3, 2, 2])
         with self.assertRaises(TypeError):
             params.read_parameters(3)
 
     def test_read_parameters_filename_default_existance(self):
-        params = ffdp.FFDParameters(n_control_points=[3, 2, 2])
+        params = FFDParameters(n_control_points=[3, 2, 2])
         params.read_parameters()
         outfilename = 'parameters.prm'
         assert os.path.isfile(outfilename)
         os.remove(outfilename)
 
     def test_read_parameters_filename_default(self):
-        params = ffdp.FFDParameters(n_control_points=[3, 2, 2])
+        params = FFDParameters(n_control_points=[3, 2, 2])
         params.read_parameters()
         outfilename = 'parameters.prm'
         outfilename_expected = 'tests/test_datasets/parameters_default.prm'
@@ -235,19 +235,19 @@ class TestFFDParameters(TestCase):
         os.remove(outfilename)
 
     def test_write_parameters_failing_filename_type(self):
-        params = ffdp.FFDParameters(n_control_points=[3, 2, 2])
+        params = FFDParameters(n_control_points=[3, 2, 2])
         with self.assertRaises(TypeError):
             params.write_parameters(5)
 
     def test_write_parameters_filename_default_existance(self):
-        params = ffdp.FFDParameters(n_control_points=[3, 2, 2])
+        params = FFDParameters(n_control_points=[3, 2, 2])
         params.write_parameters()
         outfilename = 'parameters.prm'
         assert os.path.isfile(outfilename)
         os.remove(outfilename)
 
     def test_write_parameters_filename_default(self):
-        params = ffdp.FFDParameters(n_control_points=[3, 2, 2])
+        params = FFDParameters(n_control_points=[3, 2, 2])
         params.write_parameters()
         outfilename = 'parameters.prm'
         outfilename_expected = 'tests/test_datasets/parameters_default.prm'
@@ -256,7 +256,7 @@ class TestFFDParameters(TestCase):
         os.remove(outfilename)
 
     def test_write_parameters(self):
-        params = ffdp.FFDParameters(n_control_points=[3, 2, 2])
+        params = FFDParameters(n_control_points=[3, 2, 2])
         params.read_parameters('tests/test_datasets/parameters_sphere.prm')
 
         outfilename = 'tests/test_datasets/parameters_sphere_out.prm'
@@ -267,14 +267,14 @@ class TestFFDParameters(TestCase):
         os.remove(outfilename)
 
     def test_print(self):
-        params = ffdp.FFDParameters(n_control_points=[3, 2, 2])
+        params = FFDParameters(n_control_points=[3, 2, 2])
         print(params)
 
     def test_build_bounding_box_1(self):
         origin = np.array([0., 0., 0.])
         tops = np.array([1., 1., 1.])
         cube = BRepPrimAPI_MakeBox(*tops).Shape()
-        params = ffdp.FFDParameters()
+        params = FFDParameters()
         params.build_bounding_box(cube)
 
         np.testing.assert_array_almost_equal(params.lenght_box, tops, decimal=5)
@@ -283,7 +283,7 @@ class TestFFDParameters(TestCase):
         origin = np.array([0., 0., 0.])
         tops = np.array([1., 1., 1.])
         cube = BRepPrimAPI_MakeBox(*tops).Shape()
-        params = ffdp.FFDParameters()
+        params = FFDParameters()
         params.build_bounding_box(cube)
 
         np.testing.assert_almost_equal(
@@ -295,42 +295,23 @@ class TestFFDParameters(TestCase):
         np.testing.assert_almost_equal(
             params.position_vertex_3, [0., 0., 1.], decimal=5)
 
-    def test_set_box_dimension(self):
-        origin = np.array([0., 0., 0.])
-        tops = np.array([10., 10., 10.])
-        params = ffdp.FFDParameters()
-        params.origin_box = origin
-        params._set_box_dimensions(origin, tops)
-        np.testing.assert_array_almost_equal(params.lenght_box, tops)
-
     def test_set_position_of_vertices(self):
         vertex_0 = [0., 0., 0.]
         vertex_1 = [1., 0., 0.]
         vertex_2 = [0., 1., 0.]
         vertex_3 = [0., 0., 1.]
         tops = np.array([1., 1., 1.])
-        params = ffdp.FFDParameters()
+        params = FFDParameters()
         params.origin_box = vertex_0
-        params._set_box_dimensions(vertex_0, tops)
+        params.lenght_box = tops - vertex_0
         params._set_position_of_vertices()
         np.testing.assert_equal(params.position_vertex_0, vertex_0)
         np.testing.assert_equal(params.position_vertex_1, vertex_1)
         np.testing.assert_equal(params.position_vertex_2, vertex_2)
         np.testing.assert_equal(params.position_vertex_3, vertex_3)
 
-    def test_set_mapping(self):
-        origin = np.array([0., 0., 0.])
-        tops = np.array([10., 10., 10.])
-        params = ffdp.FFDParameters()
-        params.origin_box = origin
-        params._set_box_dimensions(origin, tops)
-        params._set_mapping()
-        for i in range(3):
-            self.assertEqual(params.psi_mapping[i][i], 1. / tops[i])
-            self.assertEqual(params.inv_psi_mapping[i][i], tops[i])
-
     def test_set_modification_parameters_to_zero(self):
-        params = ffdp.FFDParameters([5, 5, 5])
+        params = FFDParameters([5, 5, 5])
         params._set_transformation_params_to_zero()
         np.testing.assert_almost_equal(
             params.array_mu_x, np.zeros(shape=(5, 5, 5)))
@@ -343,7 +324,7 @@ class TestFFDParameters(TestCase):
         min_vals = np.zeros(3)
         max_vals = np.ones(3)
         cube = BRepPrimAPI_MakeBox(1, 1, 1).Shape()
-        params = ffdp.FFDParameters()
+        params = FFDParameters()
         xyz_min, xyz_max = params._calculate_bb_dimension(cube)
         np.testing.assert_almost_equal(xyz_min, min_vals, decimal=5)
         np.testing.assert_almost_equal(xyz_max, max_vals, decimal=5)
@@ -355,7 +336,7 @@ class TestFFDParameters(TestCase):
         box = BRepPrimAPI_MakeBox(a, b).Shape()
         sphere = BRepPrimAPI_MakeSphere(3).Shape()
         section = BRepAlgoAPI_Cut(box, sphere).Shape()
-        params = ffdp.FFDParameters()
+        params = FFDParameters()
         xyz_min, xyz_max = params._calculate_bb_dimension(
             section, triangulate=True)
         correct_min = -1 * np.ones(3)
