+++ tests/test_rbfparams.py
@@ -1,84 +1,89 @@
 from unittest import TestCase
 import unittest
-import pygem.params as rbfp
 import numpy as np
 import filecmp
 import os
 
+from pygem.params_rbf import RBFParameters
+
+unit_cube = np.array([
+    [0., 0., 0.],
+    [0., 0., 1.],
+    [0., 1., 0.],
+    [1., 0., 0.],
+    [0., 1., 1.],
+    [1., 0., 1.],
+    [1., 1., 0.],
+    [1., 1., 1.]])
 
 class TestRBFParameters(TestCase):
     def test_class_members_default_basis(self):
-        params = rbfp.RBFParameters()
-        assert params.basis == None
+        params = RBFParameters()
+        assert params.basis == 'gaussian_spline'
 
     def test_class_members_default_radius(self):
-        params = rbfp.RBFParameters()
-        assert params.radius == None
+        params = RBFParameters()
+        assert params.radius == 0.5
 
     def test_class_members_default_n_control_points(self):
-        params = rbfp.RBFParameters()
-        assert params.n_control_points == None
+        params = RBFParameters()
+        assert params.n_control_points == 8
 
     def test_class_members_default_original_control_points(self):
-        params = rbfp.RBFParameters()
-        assert params.original_control_points == None
+        params = RBFParameters()
+        np.testing.assert_array_equal(params.original_control_points, unit_cube)
 
     def test_class_members_default_deformed_control_points(self):
-        params = rbfp.RBFParameters()
-        assert params.deformed_control_points == None
+        params = RBFParameters()
+        np.testing.assert_array_equal(params.deformed_control_points, unit_cube)
 
     def test_read_parameters_basis(self):
-        params = rbfp.RBFParameters()
+        params = RBFParameters()
         params.read_parameters('tests/test_datasets/parameters_rbf_default.prm')
         assert params.basis == 'gaussian_spline'
 
     def test_read_parameters_radius(self):
-        params = rbfp.RBFParameters()
+        params = RBFParameters()
         params.read_parameters('tests/test_datasets/parameters_rbf_default.prm')
         assert params.radius == 0.5
 
     def test_read_parameters_n_control_points(self):
-        params = rbfp.RBFParameters()
+        params = RBFParameters()
         params.read_parameters('tests/test_datasets/parameters_rbf_default.prm')
         assert params.n_control_points == 8
 
     def test_read_parameters_original_control_points(self):
-        params = rbfp.RBFParameters()
+        params = RBFParameters()
         params.read_parameters('tests/test_datasets/parameters_rbf_default.prm')
-        original_control_points_exact = np.array([0., 0., 0., 0., 0., 1., 0., 1., 0., 1., 0., 0., \
-          0., 1., 1., 1., 0., 1., 1., 1., 0., 1., 1., 1.]).reshape((8, 3))
         np.testing.assert_array_almost_equal(params.original_control_points,
-                                             original_control_points_exact)
+                                             unit_cube)
 
     def test_read_parameters_deformed_control_points(self):
-        params = rbfp.RBFParameters()
+        params = RBFParameters()
         params.read_parameters('tests/test_datasets/parameters_rbf_default.prm')
-        deformed_control_points_exact = np.array([0., 0., 0., 0., 0., 1., 0., 1., 0., 1., 0., 0., \
-          0., 1., 1., 1., 0., 1., 1., 1., 0., 1., 1., 1.]).reshape((8, 3))
         np.testing.assert_array_almost_equal(params.deformed_control_points,
-                                             deformed_control_points_exact)
+                                             unit_cube)
 
     def test_read_parameters_failing_filename_type(self):
-        params = rbfp.RBFParameters()
+        params = RBFParameters()
         with self.assertRaises(TypeError):
             params.read_parameters(3)
 
     def test_read_parameters_failing_number_deformed_control_points(self):
-        params = rbfp.RBFParameters()
+        params = RBFParameters()
         with self.assertRaises(TypeError):
             params.read_parameters(
                 'tests/test_datasets/parameters_rbf_bugged_01.prm')
 
     def test_write_parameters_failing_filename_type(self):
-        params = rbfp.RBFParameters()
+        params = RBFParameters()
         with self.assertRaises(TypeError):
             params.write_parameters(5)
 
     def test_write_parameters_filename_default_existance(self):
-        params = rbfp.RBFParameters()
+        params = RBFParameters()
         params.basis = 'inv_multi_quadratic_biharmonic_spline'
         params.radius = 0.1
-        params.n_control_points = 3
         params.original_control_points = np.array(
             [0., 0., 0., 0., 0., 1., 0., 1., 0.]).reshape((3, 3))
         params.deformed_control_points = np.array(
@@ -89,15 +94,12 @@ class TestRBFParameters(TestCase):
         os.remove(outfilename)
 
     def test_write_parameters_filename_default(self):
-        params = rbfp.RBFParameters()
+        params = RBFParameters()
         params.basis = 'gaussian_spline'
         params.radius = 0.5
-        params.n_control_points = 8
         params.power = 2
-        params.original_control_points = np.array([0., 0., 0., 0., 0., 1., 0., 1., 0., 1., 0., 0., \
-         0., 1., 1., 1., 0., 1., 1., 1., 0., 1., 1., 1.]).reshape((8, 3))
-        params.deformed_control_points = np.array([0., 0., 0., 0., 0., 1., 0., 1., 0., 1., 0., 0., \
-         0., 1., 1., 1., 0., 1., 1., 1., 0., 1., 1., 1.]).reshape((8, 3))
+        params.original_control_points = unit_cube
+        params.deformed_control_points = unit_cube
         outfilename = 'test.prm'
         params.write_parameters(outfilename)
         outfilename_expected = 'tests/test_datasets/parameters_rbf_default.prm'
@@ -107,7 +109,7 @@ class TestRBFParameters(TestCase):
         os.remove(outfilename)
 
     def test_write_parameters(self):
-        params = rbfp.RBFParameters()
+        params = RBFParameters()
         params.read_parameters('tests/test_datasets/parameters_rbf_cube.prm')
 
         outfilename = 'tests/test_datasets/parameters_rbf_cube_out.prm'
@@ -118,7 +120,7 @@ class TestRBFParameters(TestCase):
         os.remove(outfilename)
 
     def test_read_parameters_filename_default(self):
-        params = rbfp.RBFParameters()
+        params = RBFParameters()
         params.read_parameters()
         outfilename = 'parameters_rbf.prm'
         outfilename_expected = 'tests/test_datasets/parameters_rbf_default.prm'
@@ -127,5 +129,5 @@ class TestRBFParameters(TestCase):
         os.remove(outfilename)
 
     def test_print_info(self):
-        params = rbfp.RBFParameters()
+        params = RBFParameters()
         print(params)
